Item クラスをファイルにする
